pyproject.toml
[project]
name = "pke"
version = "0.1.0"
description = "Personal Knowledge Engine — RAG system backed by Qdrant + Ollama"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "qdrant-client>=1.12.0",
    "ollama>=0.4.0",
    "langchain-text-splitters>=0.3.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.6.0",
]

[project.scripts]
pke = "pke.cli.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
